food_algo.py
import cv2
import numpy as np
import tensorflow as tf
from tensorflow.keras.applications import MobileNetV2
from tensorflow.keras.layers import Dense, GlobalAveragePooling2D
from tensorflow.keras.models import Model

# Step 1: Define utility functions
def preprocess_image(image):
    image = cv2.resize(image, (224, 224))
    image = image / 255.0
    return image

def display_result(frame, label):
    cv2.putText(frame, label, (10, 30), cv2.FONT_HERSHEY_SIMPLEX, 1, (255, 0, 0), 2)

# Step 2: Collect and preprocess data
# This step involves preparing your dataset which is not shown here for brevity

# Step 3: Train the model (pseudo-code, assuming you have the dataset ready)
def train_model(num_classes):
    base_model = MobileNetV2(weights='imagenet', include_top=False)
    x = base_model.output
    x = GlobalAveragePooling2D()(x)
    x = Dense(1024, activation='relu')(x)
    predictions = Dense(num_classes, activation='softmax')(x)
    model = Model(inputs=base_model.input, outputs=predictions)
    model.compile(optimizer='adam', loss='categorical_crossentropy', metrics=['accuracy'])
    # Assuming you have prepared train_generator
    model.fit(train_generator, epochs=10, steps_per_epoch=train_generator.samples // train_generator.batch_size)
    model.save('food_detection_model.h5')

# Step 4: Convert the model to .pb file
def convert_to_pb():
    model = tf.keras.models.load_model('food_detection_model.h5')
    tf.saved_model.save(model, 'saved_model')
    converter = tf.lite.TFLiteConverter.from_saved_model('saved_model')
    tflite_model = converter.convert()
    with open('food_detection_model.tflite', 'wb') as f:
        f.write(tflite_model)

# Step 5: Set up live camera detection
def live_camera_detection():
    model = tf.saved_model.load('saved_model')
    class_names = ['class1', 'class2', 'class3']  # Replace with actual class names
    cap = cv2.VideoCapture(0)

    while True:
        ret, frame = cap.read()
        if not ret:
            break

        image = preprocess_image(frame)
        input_tensor = tf.convert_to_tensor([image], dtype=tf.float32)
        detections = model(input_tensor)
        label = class_names[np.argmax(detections)]

        display_result(frame, label)
        cv2.imshow('Food Detection', frame)

        if cv2.waitKey(1) & 0xFF == ord('q'):
            break

    cap.release()
    cv2.destroyAllWindows()

# Execute the steps
# 1. train_model(num_classes) # Uncomment to train the model
# 2. convert_to_pb()          # Uncomment to convert the model
# 3. live_camera_detection()  # Uncomment to run live detection
# Real-Time Object Detection