requirements.txt
numpy==1.23.5
pandas==1.5.3
matplotlib==3.6.2
seaborn==0.12.1
scikit-learn==1.1.3
tensorflow==2.11.0  # or keras if you prefer
opencv-python==4.6.0.66
Pillow==9.3.0
nltk==3.7
spacy==3.4.4
torch==1.13.0
torchvision==0.14.0
jupyterlab==3.4.8
notebook==6.5.2
scipy==1.9.3
Optional: for working with GPU (CUDA)
tensorflow-gpu==2.11.0
torch==1.13.0+cu116
scikit-optimize==0.9.0
beautifulsoup4==4.11.1
requests==2.28.1
